Python/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
greenlet==1.1.2
idna==3.3
protobuf==3.19.1
pycryptodome==3.14.1
requests==2.27.1
SQLAlchemy==1.4.36
urllib3==1.26.8
